# c2c-e-commerce/src/lib/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "C2C E-Commerce API",
    "version": "1.0.0",
    "description": "RESTful API for the C2C (consumer-to-consumer) e-commerce marketplace. Supports user authentication, listings management, orders, reviews, and categories.",
    "contact": {
      "name": "C2C Market"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the JWT token obtained from POST /api/auth/login"
      }
    },
...
  },
...
  "paths": {
...
    "/api/images/{id}": {
      "get": {
        "tags": [
          "Listings"
        ],
        "summary": "Fetch a listing photo",
        "description": "Publicly visible for a published listing — `active`, `reserved` or `sold` — matching when the listings API itself would show the parent listing to an anonymous caller, plus `reserved` and `sold` so a pending or completed purchase's photos keep rendering. A `draft` or `removed` listing's photos are visible only to its owner or an admin. Streams the stored bytes same-origin. Storage keys are random, so a URL never outlives the bytes it names.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The image",
            "content": {
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid image id"
          },
          "404": {
            "description": "Image not found, or not visible to this caller"
          }
        }
      }
    },
...
  }
}